Fix bugs

+++ jetrov_control/include/jetrov_control/common.h
@@ -40,7 +40,7 @@ int getkey()
     return character;
 }
 
-int map(int x, int in_min, int in_max, int out_min, int out_max)
+int map(double x, int in_min, int in_max, int out_min, int out_max)
 {
     int toReturn = (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min ;
